@@ -1,31 +1,29 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
     <LangVersion>latest</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="EMG.Extensions.AspNetCore.Authentication.JWT" Version="1.1.0" />
+    <PackageReference Include="EMG.Extensions.AspNetCore.Authentication.JWT" Version="1.1.1" />
     <PackageReference Include="EMG.Extensions.AspNetCore.Filters.Error" Version="1.1.0" />
-    <PackageReference Include="EMG.Extensions.Logging.Loggly" Version="1.0.2" />
+    <PackageReference Include="EMG.Extensions.Logging.Loggly" Version="1.1.0" />
     <PackageReference Include="Kralizek.Extensions.Logging" Version="2.0.1" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
 <!--#if (AddWcfDiscovery) -->
-    <PackageReference Include="EMG.Wcf.Discovery" Version="1.0.25" />
+    <PackageReference Include="EMG.Extensions.DependencyInjection.ServiceModel.DiscoveryAdapter" Version="1.1.0" />
 <!--#endif-->
 <!--#if (AddNybus) -->
-    <PackageReference Include="Nybus.Engine.RabbitMq" Version="1.5.1" />
-    <PackageReference Include="Nybus" Version="1.5.1" />
-    <PackageReference Include="Nybus.Extensions.Hosting.HostedService" Version="1.5.1" />
+    <PackageReference Include="Nybus.Engine.RabbitMq" Version="1.5.3" />
+    <PackageReference Include="Nybus" Version="1.5.3" />
+    <PackageReference Include="Nybus.Extensions.Hosting.HostedService" Version="1.5.3" />
 <!--#endif -->
 <!--#if (AddNybusBridge || ConfigureAWS) -->
-    <PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.3.100.1" />
+    <PackageReference Include="AWSSDK.Extensions.NETCore.Setup" Version="3.3.101" />
 <!--#endif-->
 <!--#if (AddNybusBridge)-->
-    <PackageReference Include="EMG.Common.Framework.NybusBridge" Version="2.4.0" />
+    <PackageReference Include="EMG.Utilities.NybusBridge" Version="1.9.0" />
 <!--#endif -->
   </ItemGroup>
 
